Fix Ubuntu build

+++ Baikal/Utils/cl_mipmap_generator.cpp
@@ -109,7 +109,7 @@ namespace Baikal
         {
             bool is_rounding_necessary = (src_size.x % 2 != 0);
 
-            if (m_x_weights.GetElementCount() < dst_size.x)
+            if (static_cast<int>(m_x_weights.GetElementCount()) < dst_size.x)
             {
                 m_x_weights = GetContext().CreateBuffer<float4>(dst_size.x, CL_MEM_READ_WRITE);
             }
@@ -137,7 +137,7 @@ namespace Baikal
         {
             bool is_rounding_necessary = (src_size.y % 2 != 0);
 
-            if (m_y_weights.GetElementCount() < dst_size.y)
+            if (static_cast<int>(m_y_weights.GetElementCount()) < dst_size.y)
             {
                 m_y_weights = GetContext().CreateBuffer<RadeonRays::float4>(dst_size.y, CL_MEM_READ_WRITE);
             }
@@ -192,3 +192,4 @@ namespace Baikal
         return std::make_shared<CLMipmapGeneratorConcrete>(context, program_manager);
     }
 }
+
